requirements.txt
autopep8==2.0.2
boto3==1.26.123
botocore==1.29.123
click==8.1.3
configparser==5.3.0
iniconfig==2.0.0
jmespath==1.0.1
markdown-it-py==2.2.0
mdurl==0.1.2
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
Pygments==2.15.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
rich==13.3.4
s3transfer==0.6.0
sh==2.0.3
six==1.16.0
tomli==2.0.1
typer==0.7.0
urllib3==1.26.15
